+++ server/views/index.jade
@@ -1,22 +1,13 @@
-extends layout
-
-block content
-  h1= title
-  p Welcome to #{title}
-  if(user)
+if(user)
     ul
       li
         img(src="#{user.photo}")
       li Bienvenido #{user.name}
       li
         a(href='logout') Salir
-  else
+else
     ul
       li
         a(href='auth/twitter') Login con Twitter
       li
         a(href='auth/facebook') Login con Facebook
-      li
-        a(href='auth/google') Login con Google
-
-
