--- server/views/layout.jade
@@ -1,7 +1,0 @@
-doctype
-html
-  head
-    title= title
-    link(rel='stylesheet', href='/stylesheets/style.css')
-  body
-    block content
